Add network file: manage upnp PAT
