+++ agent.py
@@ -1,55 +1,17 @@
-import socket
+
+
 import threading
 import argparse
 import pyfiglet
-import sys
+
 
 from utils.user import generate_username
+from utils.network import get_local_ip, mapping_port, init_upnp, check_mapping, listener, sender
+
 
 def print_sileo():
     ascii_art = pyfiglet.figlet_format("Sileo", font="slant")
     print(ascii_art)
-
-
-def listener(host:str, port:int, username:str):
-    server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    server.bind((host, port))
-    server.listen(1)
-    print(f"En attente de connexion sur {host}:{port}...")
-    conn, addr = server.accept()
-    print(f"Connexion établie avec {addr}")
-
-    while True:
-        try:
-            data = conn.recv(1024)
-            if not data:
-                break
-            message = data.decode('utf-8')
-            # Efface la ligne de saisie utilisateur
-            sys.stdout.write('\r' + ' ' * 80 + '\r')
-            sys.stdout.write(f"AnonymeUser >> {message}\n")
-            sys.stdout.write(f"{username}(you) >> ")
-            sys.stdout.flush()
-        except ConnectionResetError:
-            break
-
-def sender(target_host:str, target_port:int, username:str):
-    client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    client.connect((target_host, target_port))
-    print(f"Connecté à {target_host}:{target_port}")
-
-    while True:
-        msg = input(f"{username}(you)>> ")
-        if msg.lower() == "exit":
-            break
-        client.send(msg.encode('utf-8'))
-    client.close()
-
-def get_local_ip():
-    hostname = socket.gethostname()
-    local_host = socket.gethostbyname(hostname)
-    print(f"Your local adresse IP: {local_host}")
-    return local_host
 
 
 def main(target_ip):
@@ -61,13 +23,23 @@ def main(target_ip):
     #local_host = '172.30.160.1'
     recv_port = 1501
 
+    #initialisation de l'upnp
+    upnp = init_upnp()
+    
+    #Creation du PAT, par defaut -> port intern: 1501, port extern: 32245
+    mapping_port(upnp)
+    
+    #Check du mapping
+    check_mapping(upnp)
+    #faire un test plus tard
+    
     # Lancer l'écouteur dans un thread
     threading.Thread(target=listener, args=(local_host, recv_port, username), daemon=True).start()
 
     # Envoi des messages
     target_host = '172.30.160.68'
-    target_port = 1501         # port distant d'écoute
-    sender(target_ip, target_port, username) #Se connecte à la machine distante sur le port 1501
+    target_port = 32245         # port distant d'écoute
+    sender(target_ip, target_port, username,upnp) #Se connecte à la machine distante sur le port 1501
     
 if __name__ == "__main__":
