+++ utils/__init__.py
@@ -1,1 +1,2 @@
 from .user import generate_username
+from .network import get_local_ip, mapping_port, init_upnp, delete_mapping, check_mapping
